metodos json

+++ src/Products/Clothing.java
@@ -10,12 +10,12 @@ public class Clothing extends Products{
 		
 	}
 
-	public Clothing(int id, String name, float price, int stock) {
+	public Clothing(int id, String name, double price, int stock) {
 		super(id, name, price, stock);
 	}
 
 
-	public Clothing(int id, String name, float price, int stock, char size, String colour, String brand) {
+	public Clothing(int id, String name, double price, int stock, char size, String colour, String brand) {
 		super(id, name, price, stock);
 		this.size = size;
 		this.colour = colour;
